Use setlocale. Catch errors when setting locale. Restore environment after test.
For me the test didn't run without also calling setlocale(). Added
this based on the PHPUnit setLocale() call which resets to the
old locale after testing.

In case setting the locale fails the tests will be skipped.

I'm not certain how environment variables will be handled by PHPUnit
but to be on the safe side I rather reset them after the test.